feat(modules)!: give every helper exactly one home
Executes the decisions recorded in MODULE-BOUNDARIES.md, which #174 opened
to settle. Thirteen names were exported from two modules each and five more
were the same function under two names; the losing copy is deleted rather
than deprecated, so a consumer gets a build error naming the fix instead of
a silent behaviour change. Beta is the only window where that is free.

`./formatting` and `./math` are removed — 46 subpaths become 44. `formatting`
was the largest source of ambiguity because it named an activity rather than
a subject, so every module that formats anything overlapped with it. `math`
was four one-line wrappers around the arithmetic operators.

Three of the surviving copies were the stricter of the pair, and
`html.unescapeHtml` adopts the `strings` implementation, which decoded three
entities the `html` one missed. The before/after map for consumers is in the
migration guide.

`./logger` stays separate from `./logging` — that split is a dependency
boundary (pino vs zero runtime imports), not a naming accident.

Closes #174

BREAKING CHANGE: `./formatting` and `./math` are removed. Where two
modules exported the same name, only the owner keeps it: `roundTo`
(numbers, gone from currency), `randomString` (random, gone from
strings), `sanitizeString` (security, gone from strings), `escapeHtml`
and `unescapeHtml` (html, gone from strings), `pluralize` (strings,
gone from i18n), `secondsBetween` (time, gone from datetime),
`isBoolean` (validation, gone from boolean), and `getProcessUptime`
(process, gone from node). `events.once` is renamed `onceEvent`;
`numbers.getRandomInt` and `getRandomFloat` are `random.randomInt`
and `randomFloat`; `strings.stripHtml` is `html.stripHtmlTags`;
`formatting.formatPercent` is now `numbers.formatPercent`. Note that
`formatting.formatDate` was local time while `date.formatDate` is UTC.
