+++ apps/docs/docs/index.mdx
@@ -81,7 +81,7 @@ See [Tree-shaking & imports](./guides/tree-shaking.md) for the full explanation.
 | --- | --- | --- |
 | Dates & time | `date`, `datetime`, `time` | `today`, `daysBetween`, `formatRelative`, `nowIso`, `unixTimestamp` |
 | Numbers & math | `numbers`, `math`, `currency` | `sum`, `average`, `roundTo`, `clamp`, `formatCurrency` |
-| Strings | `strings`, `formatting`, `regex` | `slugify`, `truncate`, `capitalize`, `removeEmojis` |
+| Strings | `strings`, `html`, `regex` | `slugify`, `truncate`, `capitalize`, `removeEmojis` |
 | Collections | `arrays`, `objects`, `sets`, `maps` | `unique`, `chunk`, `groupBy`, `deepMerge`, `pick`, `omit` |
 | Async | `promises`, `sleep`, `abortController`, `functions` | `retry`, `timeout`, `delay`, `debounce`, `throttle` |
 | Validation | `validation`, `emails`, `url`, `uuid` | `isValidEmail`, `isValidUrl`, `generateUUID` |
