+++ apps/docs/docs/modules/boolean.md
@@ -3,10 +3,10 @@ title: Boolean
 description: Utilities exported from @rtorcato/js-common/boolean.
 ---
 
-Named logical operators (`and`, `or`, `not`, `xor`) plus a boolean type guard and a coercion helper. The operators exist so boolean logic can be passed around as a value — into `reduce`, `map` or a `pipe` — not to replace `&&` and `||` in ordinary conditions. `toBoolean` understands the string forms `'true'`, `'false'`, `'1'` and `'0'` that `Boolean()` gets wrong for env vars and query params.
+Named logical operators (`and`, `or`, `not`, `xor`) plus a coercion helper. The operators exist so boolean logic can be passed around as a value — into `reduce`, `map` or a `pipe` — not to replace `&&` and `||` in ordinary conditions. `toBoolean` understands the string forms `'true'`, `'false'`, `'1'` and `'0'` that `Boolean()` gets wrong for env vars and query params.
 
 ```ts
-import { and, isBoolean, not } from '@rtorcato/js-common/boolean'
+import { and, not, toBoolean } from '@rtorcato/js-common/boolean'
 ```
 
 ## Exports
@@ -14,7 +14,6 @@ import { and, isBoolean, not } from '@rtorcato/js-common/boolean'
 | Name | Summary |
 | --- | --- |
 | `and` | Logical AND for two booleans. |
-| `isBoolean` | Checks if a value is a boolean. |
 | `not` | Logical NOT for a boolean. |
 | `or` | Logical OR for two booleans. |
 | `toBoolean` | Converts a value to a boolean. |
@@ -22,6 +21,6 @@ import { and, isBoolean, not } from '@rtorcato/js-common/boolean'
 
 ## See also
 
-- [validation](./validation.md) — type guards — `isString`, `isNumber`, `isUrl`
+- [validation](./validation.md) — type guards, including `isBoolean`
 - [numbers](./numbers.md) — sum, average, clamp, roundTo
-- [math](./math.md) — add, subtract, multiply, divide
+- [strings](./strings.md) — slugify, truncate, casing, emoji stripping
