+++ apps/docs/docs/modules/events.md
@@ -6,7 +6,7 @@ description: Utilities exported from @rtorcato/js-common/events.
 Thin wrappers over the DOM `EventTarget` API — add a listener, dispatch a `CustomEvent`, await a single event as a promise. `on` returns its own removal function, so cleanup is a value you can store and call rather than a `removeEventListener` you must remember to mirror argument for argument. These are browser-side helpers; Node's `EventEmitter` is a different API and is not covered here.
 
 ```ts
-import { emit, on, once } from '@rtorcato/js-common/events'
+import { emit, on, onceEvent } from '@rtorcato/js-common/events'
 ```
 
 ## Exports
@@ -15,12 +15,12 @@ import { emit, on, once } from '@rtorcato/js-common/events'
 | --- | --- |
 | `emit` | Dispatches a custom event on the target. |
 | `on` | Adds an event listener and returns a function to remove it. |
-| `once` | Waits for a single event to occur and resolves a promise. |
+| `onceEvent` | Waits for a single event to occur and resolves a promise. |
 | `preventDefault` | Prevents the default action for an event. |
 | `stopPropagation` | Stops propagation for an event. |
 
 ## See also
 
 - [abortController](./abortController.md) — cancel in-flight work with an `AbortSignal`
-- [functions](./functions.md) — debounce, throttle, once, compose
+- [functions](./functions.md) — debounce, throttle, `once` (memoise a call), compose
 - [promises](./promises.md) — delay, timeout and settle helpers
