+++ apps/docs/docs/modules/html.md
@@ -17,11 +17,11 @@ import { containsHtml, escapeHtml, stripHtmlTags } from '@rtorcato/js-common/htm
 | `escapeHtml` | Escapes special HTML characters in a string to prevent XSS attacks. |
 | `stripHtmlTags` | Strips all HTML tags from a string. |
 | `textToHtml` | Converts a plain text string to a simple HTML paragraph (\n to ). |
-| `unescapeHtml` | Unescapes HTML entities in a string. |
+| `unescapeHtml` | Unescapes HTML entities in a string, including `&#x27;`, `&#x2F;` and `&nbsp;`. |
 
 ## See also
 
 - [strings](./strings.md) — slugify, truncate, casing, emoji stripping
 - [security](./security.md) — password strength, secure tokens, sanitizing
 - [regex](./regex.md) — escape patterns, match-all, split by regex
-- [formatting](./formatting.md) — padding, thousands separators, percentages
+- [validation](./validation.md) — type guards — `isString`, `isNumber`, `isUrl`
