+++ apps/docs/docs/modules/i18n.md
@@ -3,7 +3,7 @@ title: I18n
 description: Utilities exported from @rtorcato/js-common/i18n.
 ---
 
-A minimal localisation layer — locale detection, `Intl`-backed number and date formatting, and a dictionary-lookup `t`. Formatting delegates to `Intl`, so the locale data is the platform's rather than the bundle's, while translation and `pluralize` are deliberately naive: a flat dictionary and basic English plurals. For ICU messages, per-language plural categories or namespaced catalogs, use `i18next` or FormatJS.
+A minimal localisation layer — locale detection, `Intl`-backed number and date formatting, and a dictionary-lookup `t`. Formatting delegates to `Intl`, so the locale data is the platform's rather than the bundle's, while translation is deliberately naive: a flat dictionary lookup. For ICU messages, per-language plural categories or namespaced catalogs, use `i18next` or FormatJS. English-only pluralisation lives in [strings](./strings.md).
 
 ```ts
 import { detectLanguage, formatDateI18n, formatNumber } from '@rtorcato/js-common/i18n'
@@ -16,12 +16,11 @@ import { detectLanguage, formatDateI18n, formatNumber } from '@rtorcato/js-commo
 | `detectLanguage` | Detects the user's preferred language from the browser or Node.js environment. |
 | `formatDateI18n` | Formats a date as a localized string. |
 | `formatNumber` | Formats a number as a localized string. |
-| `pluralize` | Pluralizes a word based on a count (basic English only). |
 | `t` | Simple translation function using a dictionary object. |
 
 ## See also
 
-- [formatting](./formatting.md) — padding, thousands separators, percentages
+- [numbers](./numbers.md) — sum, average, clamp, roundTo
 - [currency](./currency.md) — price formatting, parsing and currency codes
 - [strings](./strings.md) — slugify, truncate, casing, emoji stripping
 - [date](./date.md) — calendar-day helpers — add, diff, compare, format
