+++ apps/docs/docs/modules/regex.md
@@ -24,4 +24,4 @@ import { escapeRegExp, matchAll, replaceAllRegex } from '@rtorcato/js-common/reg
 - [strings](./strings.md) — slugify, truncate, casing, emoji stripping
 - [html](./html.md) — escape, unescape and strip HTML
 - [validation](./validation.md) — type guards — `isString`, `isNumber`, `isUrl`
-- [formatting](./formatting.md) — padding, thousands separators, percentages
+- [strings](./strings.md) — slugify, truncate, casing, emoji stripping
