+++ apps/docs/docs/modules/strings.md
@@ -4,7 +4,7 @@ description: String utilities — slugify, truncate, casing.
 sidebar_position: 4
 ---
 
-Everyday string shaping — slugs, truncation, casing, emoji stripping. Every function returns a new string and leaves its input alone; `truncate` counts its ellipsis inside the requested length rather than adding to it. `slugify` strips diacritics and then keeps only `a-z0-9` and hyphens, which makes it URL-safe by construction but also means non-Latin scripts slug to an empty string — transliterate first if you need to support them.
+Everyday string shaping — slugs, truncation, casing, padding, pluralisation. Every function returns a new string and leaves its input alone; `truncate` counts its ellipsis inside the requested length rather than adding to it. `slugify` strips diacritics and then keeps only `a-z0-9` and hyphens, which makes it URL-safe by construction but also means non-Latin scripts slug to an empty string — transliterate first if you need to support them.
 
 ```ts
 import { slugify, truncate, capitalize, removeEmojis } from '@rtorcato/js-common/strings'
@@ -17,7 +17,7 @@ removeEmojis('hi 👋 there 🎉')      // "hi  there "
 
 ## See also
 
-- [formatting](./formatting.md) — padding, thousands separators, percentages
+- [i18n](./i18n.md) — locale-aware number/date formatting and translation
 - [regex](./regex.md) — escape patterns, match-all, split by regex
 - [html](./html.md) — escape, unescape and strip HTML
 - [validation](./validation.md) — type guards — `isString`, `isNumber`, `isUrl`
