+++ package.json
@@ -123,10 +123,6 @@
       "import": "./dist/file/index.js",
       "types": "./dist/file/index.d.ts"
     },
-    "./formatting": {
-      "import": "./dist/formatting/index.js",
-      "types": "./dist/formatting/index.d.ts"
-    },
     "./functions": {
       "import": "./dist/functions/index.js",
       "types": "./dist/functions/index.d.ts"
@@ -162,10 +158,6 @@
     "./maps": {
       "import": "./dist/maps/index.js",
       "types": "./dist/maps/index.d.ts"
-    },
-    "./math": {
-      "import": "./dist/math/index.js",
-      "types": "./dist/math/index.d.ts"
     },
     "./mime-types": {
       "import": "./dist/mime-types/index.js",
