+++ skills/js-common/SKILL.md
@@ -5,7 +5,7 @@ description: Use when writing TypeScript/JavaScript that needs a general-purpose
 
 # Using @rtorcato/js-common
 
-`@rtorcato/js-common` is a tree-shakeable, ESM-only utility library for Node.js >= 22 (TypeScript-first, JSDoc on every public API). One subpath export per concern — 46 of them. The package root is intentionally empty.
+`@rtorcato/js-common` is a tree-shakeable, ESM-only utility library for Node.js >= 22 (TypeScript-first, JSDoc on every public API). One subpath export per concern — 44 of them. The package root is intentionally empty.
 
 ## Rules
 
@@ -18,22 +18,16 @@ description: Use when writing TypeScript/JavaScript that needs a general-purpose
    import { slugify } from '@rtorcato/js-common'
    ```
 
-2. **The same name can exist in more than one module — pick by module, not by name.** Check this list before importing:
+2. **Every export name has exactly one home.** No name is exported from two modules, so the module → exports table below is the only lookup you need. These pairs still read alike — check which one you want:
 
-   | Name | Modules | Difference |
-   |---|---|---|
-   | `roundTo` | `numbers`, `currency` | `currency` defaults to 2 decimals |
-   | `formatDate` | `date`, `formatting` | both take a `Date`; `date` is the date-utilities home |
-   | `formatTime` | `time`, `formatting` | `time` also parses/diffs times |
-   | `formatNumber` | `formatting`, `i18n` | `i18n` is locale-driven |
-   | `once` | `functions`, `events` | `functions` = call-once wrapper; `events` = one-shot DOM listener |
-   | `escapeHtml` / `unescapeHtml` | `strings`, `html` | identical intent; `html` also has `stripHtmlTags`, `strings` has `stripHtml` |
-   | `sanitizeString` | `strings`, `security` | `security` is the one to use on untrusted input |
-   | `pluralize` | `strings`, `i18n` | `i18n` is locale-aware |
-   | `randomString` | `strings`, `random` | `strings` requires an explicit charset |
-   | `secondsBetween` | `time`, `datetime` | equivalent |
-   | `isBoolean` | `boolean`, `validation` | `validation` is a type guard |
-   | `getProcessUptime` | `node`, `process` | equivalent |
+   | Looks like | Actually |
+   |---|---|
+   | `functions.once` vs `events.onceEvent` | `once(fn)` memoises a call; `onceEvent(target, type)` awaits a DOM event |
+   | `logger` vs `logging` | `logger` is one configured pino instance; `logging` is `console` helpers with no runtime deps |
+   | `try.tryCatch` vs `errors.tryWithFallback` | Result object vs swallow-and-fall-back — see rule 3 |
+   | `date.formatDate` vs `datetime.formatDateTimeLocal` | a UTC calendar day vs a local instant |
+
+   Where two modules used to ship the same name, the loser was deleted rather than aliased — [MODULE-BOUNDARIES.md](https://github.com/rtorcato/js-common/blob/main/MODULE-BOUNDARIES.md) records which side won and why.
 
 3. **Three different error idioms — don't mix them up.**
 
@@ -67,33 +61,31 @@ Import from `@rtorcato/js-common/<module>`.
 |---|---|
 | abortController | createAbortController, abortPromise, withAbort, abortAfter |
 | arrays | first, last, unique, flatten, chunk, compact, shuffle, groupBy |
-| boolean | toBoolean, isBoolean, xor, and, or, not |
+| boolean | toBoolean, xor, and, or, not |
 | colors | randomColor, matchingTextColor, hexToRgb, rgbToHex, isValidHex, darken, lighten |
 | console | disableConsole, clearConsole |
 | crypto | hashString, randomHex, hmacHash, base64Encode, base64Decode |
-| currency | getCurrencySymbol, getCurrencyName, getCurrencyLocale, formatPrice, formatPriceCompact, parsePrice, parseCurrencyString, convertCurrency, isValidCurrencyCode, isValidCurrency, roundTo |
+| currency | getCurrencySymbol, getCurrencyName, getCurrencyLocale, formatPrice, formatPriceCompact, parsePrice, parseCurrencyString, convertCurrency, isValidCurrencyCode, isValidCurrency |
 | date | today, parseDate, formatDate, daysBetween, diffInHours, diffInMinutes, isLeapYear, addDays, subDays, addMonths, getDayOfWeek, startOfDay, endOfDay, isWeekend, isSameDay, isToday, formatRelative |
-| datetime | nowIso, parseIsoDateTime, formatDateTimeLocal, getTimezoneOffset, toUtcDate, unixTimestamp, unixMillis, secondsBetween, getIsoWeek, getIsoWeekInfo |
+| datetime | nowIso, parseIsoDateTime, formatDateTimeLocal, getTimezoneOffset, toUtcDate, unixTimestamp, unixMillis, getIsoWeek, getIsoWeekInfo |
 | emails | isValidEmail, normalizeEmail, maskEmail, getEmailDomain, isFreeEmailProvider |
 | env | getENV, isDev, isProd, isTest, getNodeEnv, checkEnv, RootApiEnvSchema |
 | errors | createCustomError, isErrorName, getErrorMessage, assert, tryWithFallback |
-| events | on, emit, once, preventDefault, stopPropagation |
+| events | on, emit, onceEvent, preventDefault, stopPropagation |
 | fetch | fetchWithTimeout, postJson, getJson, fetchText, handleApiError |
 | file | readFileAsString, writeFileAsString, fileExists, deleteFile, getFileExtension |
-| formatting | padZero, formatNumber, formatPercent, formatDate, formatTime, formatDateTime |
 | functions | once, debounce, throttle, compose, pipe |
 | geometry | distance2D, midpoint2D, angle2D, pointInRect, rectsOverlap, polygonArea |
 | html | escapeHtml, unescapeHtml, stripHtmlTags, textToHtml, containsHtml |
-| i18n | detectLanguage, formatNumber, formatDateI18n, t, pluralize |
+| i18n | detectLanguage, formatNumber, formatDateI18n, t |
 | interval | runInterval, clearIntervalById |
 | json | safeJsonParse, safeJsonStringify, isValidJson, deepCloneJson |
 | logger | logger (pino instance — pretty in dev, JSON in prod) |
 | logging | logWithTimestamp, info, warn, error, captureConsole, ConsoleLevel |
 | maps | invertMap, mapValues, mergeMaps, objectToMap, mapToObject |
-| math | add, subtract, multiply, divide |
 | mime-types | lookup, types, extensions, mimeTypes |
-| node | nodeVersionCheck, getNodeMajorVersion, isNode, requireOptional, getProcessUptime |
-| numbers | getRandomInt, getRandomFloat, clamp, roundTo, isFiniteNumber, isInteger, between, sum, average, mod, min, max |
+| node | nodeVersionCheck, getNodeMajorVersion, isNode, requireOptional |
+| numbers | clamp, roundTo, formatPercent, isFiniteNumber, isInteger, between, sum, average, mod, min, max |
 | objects | isPlainObject, deepMerge, omit, pick, deepClone |
 | os | getOsPlatform, getOsRelease, getOsArch, getHomeDir, getTmpDir |
 | process | getProcessId, getProcessUptime, getCwd, getProcessPlatform, exitProcess, isCI |
@@ -103,7 +95,7 @@ Import from `@rtorcato/js-common/<module>`.
 | security | sanitizeString, isStrongPassword, generateSecureToken |
 | sets | union, intersection, difference, isSubset, isSuperset, setToArray, arrayToSet |
 | sleep | sleep, sleepSync, sleepRandom, sleepWithAbort |
-| strings | titleCase, capitalize, camelCase, kebabCase, snakeCase, truncate, padStart, padEnd, randomString, replaceString, sanitizeString, reverse, slugify, words, wordCount, escapeHtml, unescapeHtml, stripHtml, pluralize, ordinalize, isBlank, mask, template |
+| strings | titleCase, capitalize, camelCase, kebabCase, snakeCase, truncate, padStart, padEnd, replaceString, reverse, slugify, words, wordCount, pluralize, ordinalize, isBlank, mask, template |
 | system | isMacOs, isWindows, isLinux, isIOS, isAndroid, getPlatform, isTouchDevice |
 | time | nowTime, nowTimeShort, parseTime, formatTime, secondsBetween, pad2 |
 | try | tryCatch, isSuccess, Result, Success, Failure |
